Moved files to cheerwithme package, added health endpoint and 500 on errors.
