+++ backend/build.gradle
@@ -14,7 +14,7 @@ apply plugin: 'kotlin'
 apply plugin: 'application'
 apply plugin: "com.github.johnrengelman.shadow"
 
-group 'cheerwithme'
+group 'dev.fredag.cheerwithme'
 version '0.0.1-SNAPSHOT'
 mainClassName = "io.ktor.server.netty.EngineMain"
