+++ backend/resources/application.conf
@@ -4,8 +4,10 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [ dev.fredag.ApplicationKt.module,
-            dev.fredag.FriendModuleKt.friendModule,
-            dev.fredag.InvitationModuleKt.invitationModule]
+        modules = [
+            dev.fredag.cheerwithme.ApplicationKt.module,
+            dev.fredag.cheerwithme.friends.FriendModuleKt.friendModule,
+            dev.fredag.cheerwithme.invitation.InvitationModuleKt.invitationModule
+        ]
     }
 }
