+++ backend/test/FriendModuleTest.kt
@@ -1,5 +1,6 @@
 package dev.fredag
 
+import dev.fredag.cheerwithme.friends.friendModule
 import io.ktor.http.*
 import kotlin.test.*
 import io.ktor.server.testing.*
